Refactor game state and models.
Move data models to /domain/model
Initialize default values for few properties.
Remove PointsState, DiceSetInfo and SettingsViewModelFactory as they are no longer needed.
