+++ app/src/main/java/pl/kazoroo/tavernFarkle/game/domain/model/Dice.kt
@@ -3,7 +3,9 @@ package pl.kazoroo.tavernFarkle.game.domain.model
 import pl.kazoroo.tavernFarkle.shop.domain.model.SpecialDiceName
 
 data class Dice(
-    val specialDiceName: SpecialDiceName? = null,
     val value: Int,
+    val isSelected: Boolean = false,
+    val isVisible: Boolean = true,
+    val special: SpecialDiceName? = null,
     val image: Int
 )
