--- app/src/main/java/pl/kazoroo/tavernFarkle/game/domain/model/DiceSetInfo.kt
@@ -1,7 +1,0 @@
-package pl.kazoroo.tavernFarkle.game.domain.model
-
-data class DiceSetInfo(
-    val diceList: List<Dice>,
-    val isDiceSelected: List<Boolean>,
-    val isDiceVisible: List<Boolean>
-)
