--- app/src/main/java/pl/kazoroo/tavernFarkle/game/domain/model/PointsState.kt
@@ -1,7 +1,0 @@
-package pl.kazoroo.tavernFarkle.game.domain.model
-
-data class PointsState(
-    val selectedPoints: Int,
-    val roundPoints: Int,
-    val totalPoints: Int
-)
