+++ app/src/main/java/pl/kazoroo/tavernFarkle/game/presentation/game/components/InteractiveDiceLayout.kt
@@ -43,7 +43,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import pl.kazoroo.tavernFarkle.R
-import pl.kazoroo.tavernFarkle.game.presentation.game.Dice
+import pl.kazoroo.tavernFarkle.game.domain.model.Dice
 
 @Composable
 fun InteractiveDiceLayout(
